URL Cleanup
This commit updates URLs to prefer the https protocol. Redirects are not followed to avoid accidentally expanding intentionally shortened URLs (i.e. if using a URL shortener).

# HTTP URLs that Could Not Be Fixed
These URLs were unable to be fixed. Please review them to see if they can be manually resolved.

* http://pkg.jenkins-ci.org/debian/jenkins-ci.org.key (200) could not be migrated:
   ([https](https://pkg.jenkins-ci.org/debian/jenkins-ci.org.key) result SSLHandshakeException).
* http://pkg.jenkins-ci.org/debian (301) could not be migrated:
   ([https](https://pkg.jenkins-ci.org/debian) result SSLHandshakeException).

# Fixed URLs

## Fixed But Review Recommended
These URLs were fixed, but the https status was not OK. However, the https status was the same as the http request or http redirected to an https URL, so they were migrated. Your review is recommended.

* http://repo.jenkins-ci.org/public/ (405) migrated to:
  https://repo.jenkins-ci.org/public/ ([https](https://repo.jenkins-ci.org/public/) result 405).

## Fixed Success
These URLs were switched to an https URL with a 2xx status. While the status was successful, your review is still recommended.

* http://github.com/pivotalsoftware/CIBorium migrated to:
  https://github.com/pivotalsoftware/CIBorium ([https](https://github.com/pivotalsoftware/CIBorium) result 200).
* http://opensource.org/licenses/BSD-3-Clause migrated to:
  https://opensource.org/licenses/BSD-3-Clause ([https](https://opensource.org/licenses/BSD-3-Clause) result 200).
* http://www.apache.org/licenses/LICENSE-2.0 migrated to:
  https://www.apache.org/licenses/LICENSE-2.0 ([https](https://www.apache.org/licenses/LICENSE-2.0) result 200).

# Ignored
These URLs were intentionally ignored.

* http://maven.apache.org/POM/4.0.0
* http://maven.apache.org/maven-v4_0_0.xsd
* http://www.w3.org/2001/XMLSchema-instance